backend/requirements.txt
# 📦 requirements.txt — 재고이찌 백엔드 런타임 (Python 3.11 기준)
# 목적:
# - FastAPI + SQLAlchemy 안정화
# - bcrypt 72바이트 호환
# - JWT/보안, 엑셀, 마이그레이션, Redis 포함
# - Docker + Uvicorn + Alembic 풀 스택

# ───────────────────────────────── FastAPI 스택 ─────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ───────────────────────────── DB 드라이버 / ORM ───────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.11

# ───────────────────────────── 설정 / 스키마 / 유틸 ────────────────────────────
pydantic==2.6.4
python-dotenv==1.0.1
redis==5.0.1

# (FastAPI + Pydantic v2 연동 보조)
pydantic-settings==2.2.1

# ───────────────────────────── 인증 / 보안 / 토큰 ───────────────────────────────
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# ✅ JWT 토큰 생성용 (jwt_tokens.py 에서 사용함)
PyJWT==2.9.0

# (참고용 - 필요시 JWT 구현 변경 가능)
python-jose[cryptography]==3.3.0

# ───────────────────────────── 엑셀 / 파일 ─────────────────────────────────────
openpyxl==3.1.5

# ───────────────────────────── 마이그레이션 ────────────────────────────────────
alembic==1.13.3
mako==1.3.5

# ───────────────────────────── 개발 보조 ──────────────────────────────────────
watchfiles==0.21.0
